frontend/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Flask-Login==0.6.2
dash==2.9.3
dash-bootstrap-components==1.5.0
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
gunicorn==20.1.0
plotly==5.15.0
google-cloud-storage==2.7.0

# Dependencias para Keycloak y Authlib
authlib==1.2.0